Move to standard tests file
