+++ lib/superlunary/src/test/superlunary.Tests.scala
@@ -32,8 +32,34 @@
                                                                                                   */
 package superlunary
 
+import scala.quoted.*
+
 import soundness.*
+
+import systemProperties.jre
+import temporaryDirectories.systemProperties
+
+case class Example(name: Text, count: Long)
 
 object Tests extends Suite(m"Superlunary Tests"):
   def run(): Unit =
-    ()
+    given jsonError: Tactic[JsonError] = strategies.throwUnsafely
+    given compilerError: Tactic[CompilerError] = strategies.throwUnsafely
+
+    inline given [value] => Quotes => (refs: References[Json]) => Conversion[value, Expr[value]] =
+      value =>
+        compiletime.summonInline[value is Encodable in Json].give:
+          val encoded = value.json
+          val allocation = refs.allocate(encoded)
+          '{  import strategies.throwUnsafely
+              compiletime.summonInline[value is Decodable in Json].give:
+                ${refs.array}(${Expr(allocation)}).as[value]  }
+
+    def fn(message: Example): Example = Jvm.dispatch:
+      '{  val x = ${message.name}
+          val y = ${message.count}
+          println(y)
+          Example(t"Time: $x $y ${java.lang.System.currentTimeMillis - ${message}.count}", 9)  }
+
+    println(fn(Example(t"one", java.lang.System.currentTimeMillis)))
+    println(fn(Example(t"two", java.lang.System.currentTimeMillis)))
